Moved live markets ticker autocomplete to separate file
Moved live markets ticker autocomplete to separate file
